Makefile mod ディレクトリ構造の変更

+++ ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: nfukuma <nfukuma@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/11 00:58:09 by nfukuma           #+#    #+#             */
-/*   Updated: 2022/08/11 01:02:55 by nfukuma          ###   ########.fr       */
+/*   Updated: 2022/08/11 01:08:46 by nfukuma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_pow(int num, int pow)
+int	ft_pow(int num, size_t pow)
 {
 	int	res;
